mcp_server/requirements.txt
azure-functions
PyGitHub>=2.1.1